chichito/requirements.txt
Django==5.1.2       
djangorestframework>=3.12
mysql-connector-python>=9.1.0
Pillow>=8.2.0          
django-cors-headers>=3.10.0  
djangorestframework-simplejwt>=5.2  
